teste mapeamento

+++ src/main/java/dao/GenericDAO.java
@@ -4,7 +4,7 @@ public interface GenericDAO<T> {
     public boolean insert(T obj);
     public boolean update(T obj);
     public boolean delete(T obj);
-    public T findByID(int id);
+    public T findByID(Long id);
     public java.util.List<T> listAll();
     public java.util.List<T> findByAttribute(String s);
 }
